+++ Romo/lib/Drive/RomoNetworking/RMNetworkUtilities.m
@@ -22,7 +22,7 @@
     return HEADER_SIZE;
 }
 
-+ (void)packInteger:(uint32_t)integer intoBuffer:(uint8_t *)buffer offset:(uint32_t)offset
++ (void)packInteger:(NSUInteger)integer intoBuffer:(uint8_t *)buffer offset:(uint32_t)offset
 {    
     buffer[offset++] = (integer >> 24) & 0xFF;
     buffer[offset++] = (integer >> 16) & 0xFF;
@@ -55,7 +55,7 @@
 
 - (char *)bytesWithHeader
 {
-    const uint32_t size = [self length];
+    const NSUInteger size = [self length];
     const void *data = [self bytes];
     
     char *dataWithHeader = malloc(size + HEADER_SIZE);
@@ -66,7 +66,7 @@
     return dataWithHeader;
 }
 
-- (uint32_t)sizeWithHeader
+- (NSUInteger)sizeWithHeader
 {
     return [self length] + HEADER_SIZE;
 }
